Move method 'buildTitle' to class 'Driver'
